converted more integration tests to unit test
